views: correct display and hiding of annotation text stats
