+++ app/views/annotation_categories/index.html.erb
@@ -6,14 +6,18 @@
   <%= javascript_include_tag 'MathJax_lib/MathJax.js?config=TeX-AMS-MML_HTMLorMML' %>
   <%= javascript_include_tag 'MathJax/mathjax_helpers' %>
   <script>
-    addAnnotationStatPanel = (annotation_id) => {
-      console.log(document);
-      console.log('this is running');
-      window.annotationStatPanel = window.makeAnnotationStatPanel(document.getElementById('annotation-stat-panel'),
-        {
-          annotation_id: annotation_id,
-        }
-      );
+    toggleAnnotationStatPanel = (annotation_id) => {
+      statsPanel = document.getElementById('annotation-text-' + annotation_id + '-stat-panel');
+      if (statsPanel.innerText === ''){
+        window.annotationStatPanel = window.makeAnnotationStatPanel(
+          statsPanel,
+          {
+            annotation_id: annotation_id,
+          }
+        );
+      } else {
+        window.annotationStatPanel.remove_component(statsPanel);
+      }
     }
   </script>
 <% end %>
